refactor telemetry
