+++ app/del.txt
@@ -129,5 +129,34 @@ while (data_size > 0 || eof) {
         break;
 }*/
 
+// Ping is special, since there is no component differentation - any component can ping all systems,
+// but never a specific system
+// Obsolete, rpelaced by timesync
+/*if(msg.msgid==MAVLINK_MSG_ID_PING){
+    mavlink_ping_t ping;
+    mavlink_msg_ping_decode(&msg, &ping);
+    //qDebug()<<"Got ping message sender:"<<msg.sysid<<":"<<msg.compid<<" target:"<<ping.target_system<<":"<<ping.target_component<<"seq:"<<ping.seq;
+    // We only process ping responses with our specific sys id and a matching sequence number.
+    // Note that if the user clicks the ping multiple times in rapid succession, some pings might be dropped.
+    // Check if the ping is targeted at our system
+    if(ping.seq==pingSequenceNumber && ping.target_system==QOpenHDMavlinkHelper::getSysId()){
+        const auto delta=QOpenHDMavlinkHelper::getTimeMicroseconds()-ping.time_usec;
+        const auto delta_readable=time_microseconds_readable(delta);
+        if(msg.sysid==OHD_SYS_ID_AIR){
+            AOHDSystem::instanceAir().set_last_ping_result_openhd(delta_readable.c_str());
+        }else if(msg.sysid==OHD_SYS_ID_GROUND){
+            AOHDSystem::instanceGround().set_last_ping_result_openhd(delta_readable.c_str());
+        }else{
+            qDebug()<<"Got ping from fc";
+            // almost 100% from flight controller
+            //if(msg.compid==MAV_COMP_ID_AUTOPILOT1)
+           FCMavlinkSystem::instance().set_last_ping_result_flight_ctrl(delta_readable.c_str());
+        }
+    }else{
+        //qDebug()<<"Got ping message sender:"<<msg.sysid<<":"<<msg.compid<<" target:"<<ping.target_system<<":"<<ping.target_component<<"seq:"<<ping.seq;
+    }
+    return;
+}*/
+
