+++ app/rc/openhdrc.cpp
@@ -12,7 +12,6 @@
 #include <QJoysticks.h>
 #endif
 
-#include "telemetry/mavlinktelemetry.h"
 
 #define BUFLEN 21
 #define PORT 5565 // UDP port for OpenHD RC
